Changed menu layout Created settings menu
