+++ config.xml
@@ -1,5 +1,5 @@
 <?xml version='1.0' encoding='utf-8'?>
-<widget android-versionCode="11" id="com.github.disparter.dm_shield" version="0.0.8" xmlns="http://www.w3.org/ns/widgets" xmlns:cdv="http://cordova.apache.org/ns/1.0">
+<widget android-versionCode="12" id="com.github.disparter.dm_shield" version="0.0.9" xmlns="http://www.w3.org/ns/widgets" xmlns:cdv="http://cordova.apache.org/ns/1.0">
     <name>dm-shield</name>
     <description>An app to index important information about 5e Dungeons and Dragons Edition. Like the Dungeon Master Screen.</description>
     <author email="felipe.a.b.faria@gmail.com" href="http://github.com/disparter">Felipe Faria - Disparter</author>
