+++ src/pages/settings/settings.html
@@ -1,0 +1,20 @@
+<ion-header>
+  <ion-navbar>
+    <button ion-button menuToggle>
+      <ion-icon name="menu"></ion-icon>
+    </button>
+    <ion-title>{{'MENU.SETTINGS' | translate}}</ion-title>
+  </ion-navbar>
+</ion-header>
+
+<ion-content>
+  <ion-list>
+    <ion-item-group radio-group (ionChange)="applyLanguage()" [(ngModel)]="selectedLanguage">
+      <ion-item-divider color="light">{{ 'SETTINGS.LANGUAGE' | translate }}</ion-item-divider>
+      <ion-item *ngFor="let lang of languages">
+        <ion-label>{{lang.name}}</ion-label>
+        <ion-radio [value]="lang.code"></ion-radio>
+      </ion-item>
+    </ion-item-group>
+  </ion-list>
+</ion-content>
